src/main/java/io/services/VehicleServiceImpl.java
package io.services;


import io.entity.Vehicle;
import io.exception.VehicleNotFoundException;
import io.repository.VehicleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
public class VehicleServiceImpl implements VehicleService {
    @Autowired
    private VehicleRepository vehicleRepository;

    @Transactional
    public Vehicle updateVehicle(Vehicle vehicle) {
        return vehicleRepository.save(vehicle);
    }

    @Override
    @Transactional(readOnly = true)
    public List<Vehicle> findAllVehicles() throws VehicleNotFoundException {
        return (List<Vehicle>) vehicleRepository.findAll();
    }

}
